+++ test/CacheSessionPersistenceTest.php
@@ -270,7 +270,10 @@ class CacheSessionPersistenceTest extends TestCase
         $this->assertAttributeSame('test', 'cookieName', $persistence);
 
         // These we did not
+        $this->assertAttributeSame(null, 'cookieDomain', $persistence);
         $this->assertAttributeSame('/', 'cookiePath', $persistence);
+        $this->assertAttributeSame(false, 'cookieSecure', $persistence);
+        $this->assertAttributeSame(false, 'cookieHttpOnly', $persistence);
         $this->assertAttributeSame('nocache', 'cacheLimiter', $persistence);
         $this->assertAttributeSame(10800, 'cacheExpire', $persistence);
         $this->assertAttributeNotEmpty('lastModified', $persistence);
@@ -299,12 +302,19 @@ class CacheSessionPersistenceTest extends TestCase
             '/api',
             $cacheLimiter,
             100,
-            $lastModified
+            $lastModified,
+            false,
+            'example.com',
+            true,
+            true
         );
 
         $this->assertAttributeSame($this->cachePool->reveal(), 'cache', $persistence);
         $this->assertAttributeSame('test', 'cookieName', $persistence);
         $this->assertAttributeSame('/api', 'cookiePath', $persistence);
+        $this->assertAttributeSame('example.com', 'cookieDomain', $persistence);
+        $this->assertAttributeSame(true, 'cookieSecure', $persistence);
+        $this->assertAttributeSame(true, 'cookieHttpOnly', $persistence);
         $this->assertAttributeSame($cacheLimiter, 'cacheLimiter', $persistence);
         $this->assertAttributeSame(100, 'cacheExpire', $persistence);
         $this->assertAttributeSame(
@@ -320,12 +330,21 @@ class CacheSessionPersistenceTest extends TestCase
             $this->cachePool->reveal(),
             'test',
             '/api',
-            'not-valid'
+            'not-valid',
+            100,
+            null,
+            false,
+            'example.com',
+            true,
+            true
         );
 
         $this->assertAttributeSame($this->cachePool->reveal(), 'cache', $persistence);
         $this->assertAttributeSame('test', 'cookieName', $persistence);
+        $this->assertAttributeSame('example.com', 'cookieDomain', $persistence);
         $this->assertAttributeSame('/api', 'cookiePath', $persistence);
+        $this->assertAttributeSame(true, 'cookieSecure', $persistence);
+        $this->assertAttributeSame(true, 'cookieHttpOnly', $persistence);
         $this->assertAttributeSame('nocache', 'cacheLimiter', $persistence);
     }